Add PostPolicy use in post components
Post policy allows views by any guest or user, with CRUD operations being admin-only. This is used in the Livewire component actions to add an additional layer of security.

+++ app/Livewire/Posts/PostCell.php
@@ -16,6 +16,7 @@ class PostCell extends Component
 
     public function delete(): void
     {
+        $this->authorize('delete', $this->post);
         if (!$this->post->deleteContent()) {
             $this->error("Server error writing post content to file");
             return;
